Add branch and git data to the Run ctx
