+++ internal-packages/run-engine/src/engine/systems/runAttemptSystem.ts
@@ -2,6 +2,7 @@ import { startSpan } from "@internal/tracing";
 import {
   CompleteRunAttemptResult,
   ExecutionResult,
+  GitMeta,
   StartRunAttemptResult,
   TaskRunError,
   TaskRunExecution,
@@ -33,6 +34,7 @@ import {
 import { SystemResources } from "./systems.js";
 import { WaitpointSystem } from "./waitpointSystem.js";
 import { DelayedRunSystem } from "./delayedRunSystem.js";
+import { tryCatch } from "@trigger.dev/core";
 
 export type RunAttemptSystemOptions = {
   resources: SystemResources;
@@ -280,6 +282,14 @@ export class RunAttemptSystem {
             dataType: taskRun.metadataType,
           });
 
+          let git: GitMeta | undefined = undefined;
+          if (environment.git) {
+            const parsed = GitMeta.safeParse(environment.git);
+            if (parsed.success) {
+              git = parsed.data;
+            }
+          }
+
           const execution: TaskRunExecution = {
             task: {
               id: run.lockedBy!.slug,
@@ -330,6 +340,8 @@ export class RunAttemptSystem {
               id: environment.id,
               slug: environment.slug,
               type: environment.type,
+              branchName: environment.branchName ?? undefined,
+              git,
             },
             organization: {
               id: environment.organization.id,
